pyproject.toml
[project]
name = "nx-boss"
dynamic = ["version"]
description = "Compatible server for PaperStream NX Manager"
readme = "README.md"
authors = [
    { name = "SEIAROTg", email = "seiarotg@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "ruamel-yaml>=0.19.1",
    "uvicorn>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "mypy>=1.11",
    "ruff>=0.6",
    "pre-commit>=3.8",
    "hatch>=1.12",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/nx_boss/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/nx_boss"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py314"
line-length = 100
exclude = ["src/nx_boss/_version.py"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PTH", "RUF"]
ignore = ["B008"]  # FastAPI uses function calls in defaults

[tool.mypy]
python_version = "3.14"
strict = true
ignore_missing_imports = true
